First implementation of wallabagItemSource. Remove Models/ folder.
