+++ wallabag/wallabag.Shared/ViewModel/MainViewModel.cs
@@ -4,7 +4,7 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Threading.Tasks;
 using wallabag.Common;
-using wallabag.Models;
+using wallabag.DataModel;
 using Windows.Networking.Connectivity;
 using Windows.Web.Syndication;
 
@@ -28,102 +28,15 @@ namespace wallabag.ViewModel
             get { return new ObservableCollection<ItemViewModel>(Items.Where(i => i.IsRead == true)); }
         }
 
-        private bool everythingFine
-        {
-            get
-            {
-                string wallabagUrl = AppSettings["wallabagUrl", string.Empty];
-                int userId = AppSettings["userId", 1];
-                string token = AppSettings["Token", string.Empty];
-
-                ConnectionProfile connections = NetworkInformation.GetInternetConnectionProfile();
-                bool internet = connections != null && connections.GetNetworkConnectivityLevel() == NetworkConnectivityLevel.InternetAccess;
-
-                return wallabagUrl != string.Empty && userId != 0 && token != string.Empty && internet;
-            }
-        }
-        private string buildUrl(string parameter)
-        {
-            string wallabagUrl = AppSettings["wallabagUrl", string.Empty];
-            int userId = AppSettings["userId", 1];
-            string token = AppSettings["Token", string.Empty];
-
-            if (everythingFine)
-                return string.Format("{0}?feed&type={1}&user_id={2}&token={3}", wallabagUrl, parameter, userId, token);
-
-            return string.Empty;
-        }
-
         public RelayCommand refreshCommand { get; private set; }
         private async Task RefreshItems()
         {
-            if (everythingFine)
-            {
-                StatusText = Helpers.LocalizedString("UpdatingText");
-                IsActive = true;
+            StatusText = Helpers.LocalizedString("UpdatingText");
+            IsActive = true;
 
-                Items.Clear();
+            var items = await wallabagDataSource.GetItemsAsync();
 
-                // The SyndicationClient is the class that will be used for accessing RSS feeds.
-                Windows.Web.Syndication.SyndicationClient client = new SyndicationClient();
-                string[] parameters = new string[] { "home", "fav", "archive" };
-
-                foreach (string param in parameters) // perform the following step for each of the parameters (home, fav, archive)
-                {
-                    Uri feedUri = new Uri(buildUrl(param));
-                    try
-                    {
-                        SyndicationFeed feed = await client.RetrieveFeedAsync(feedUri);
-
-                        if (feed.Items != null && feed.Items.Count > 0)
-                        {
-                            foreach (SyndicationItem item in feed.Items)
-                            {
-                                Item tmpItem = new Item();
-                                if (item.Title != null && item.Title.Text != null)
-                                {
-                                    tmpItem.Title = item.Title.Text;
-                                }
-                                if (item.Summary != null && item.Summary.Text != null)
-                                {
-                                    tmpItem.Content = item.Summary.Text;
-                                }
-                                if (item.Links != null && item.Links.Count > 0)
-                                {
-                                    tmpItem.Url = item.Links[0].Uri;
-                                }
-                                switch (param)
-                                {
-                                    // If we are in the 'fav' loop, set the IsFavourite property to 'true'.
-                                    case "fav":
-                                        tmpItem.IsFavourite = true;
-                                        break;
-
-                                    // If we are in the 'archive' loop, set the IsRead property to 'true'.
-                                    case "archive":
-                                        tmpItem.IsRead = true;
-                                        break;
-                                }
-                                // to avoid duplicate items...
-                                if (!Items.Contains(new ItemViewModel(tmpItem)))
-                                {
-                                    Items.Add(new ItemViewModel(tmpItem));
-                                }
-                            }
-                        }
-                        IsActive = false;
-
-                        // Inform the view that the item collections had changed.
-                        RaisePropertyChanged(() => unreadItems);
-                        RaisePropertyChanged(() => favouriteItems);
-                        RaisePropertyChanged(() => archivedItems);
-                    }
-                    catch
-                    {
-                        IsActive = false;
-                    }
-                }
-            }
+            IsActive = false;
         }
 
         public MainViewModel()
